requirements.txt
ultralytics
pydobot2